bindings: extract non-C-API infrastructure into headers
Split the ~5800-line duckdb_node_bindings.cpp by lifting the ~1540 lines of
helper/infrastructure code (everything before the addon class) into headers,
leaving the .cpp with just the includes and the DuckDBNodeAddon class whose 274
methods each mirror a duckdb.h C API function.

New headers in bindings/src/:
- napi_setup.h: the single include site for napi.h, preceded by its required
  NODE_ADDON_API_* #defines, so that ordering invariant lives in one place.
- bindings_config.h: the DEFAULT_DUCKDB_API config macro (used by the workers
  and the addon class).
- conversion_helpers.h: struct<->object converters + MakeValueRef.
- externals.h: CreateExternal/GetDataFromExternal templates, napi_type_tag
  constants, per-handle finalizers/wrappers, and the scalar-function holder.
- scalar_function_helpers.h: scalar-function bind/main TSFN plumbing.
- promise_workers.h: PromiseWorker base + the AsyncWorker subclasses.
- enums.h: DefineEnumMember + the Create*Enum helpers.

Each header is #pragma once and self-contained (includes napi_setup.h /
duckdb.h / its dependencies), so all seven are standalone-includable; free
functions are marked inline for ODR safety. The .cpp includes them in
dependency order before the class. binding.gyp is unchanged (only the .cpp is
compiled; headers are pulled in via #include).

Pure structural change: the extracted code and the addon class are byte-for-byte
identical to the original (modulo the added inline keyword). Native build and
bindings tests (265 passed / 3 skipped) unchanged.
