+++ bindings/src/bindings_config.h
@@ -1,0 +1,4 @@
+#pragma once
+
+// The duckdb_api config value these bindings report to DuckDB.
+#define DEFAULT_DUCKDB_API "node-neo-bindings"
