+++ bindings/src/duckdb_node_bindings.cpp
@@ -1,7 +1,4 @@
-#define NODE_ADDON_API_DISABLE_DEPRECATED
-#define NODE_ADDON_API_REQUIRE_BASIC_FINALIZERS
-#define NODE_API_NO_EXTERNAL_BUFFERS_ALLOWED
-#include "napi.h"
+#include "napi_setup.h"
 
 #include <condition_variable>
 #include <cstddef>
@@ -13,1540 +10,12 @@
 
 #include "duckdb.h"
 
-#define DEFAULT_DUCKDB_API "node-neo-bindings"
-
-// Napi helpers
-
-Napi::Reference<Napi::Value> MakeValueRef(Napi::Value value) {
-  return value.IsUndefined()
-    ? Napi::Reference<Napi::Value>() 
-    : Napi::Reference<Napi::Value>::New(value, 1);
-}
-
-// Conversion betweeen structs and objects
-
-Napi::Object MakeDateObject(Napi::Env env, duckdb_date date) {
-  auto date_obj = Napi::Object::New(env);
-  date_obj.Set("days", Napi::Number::New(env, date.days));
-  return date_obj;
-}
-
-duckdb_date GetDateFromObject(Napi::Object date_obj) {
-  auto days = date_obj.Get("days").As<Napi::Number>().Int32Value();
-  return { days };
-}
-
-Napi::Object MakeDatePartsObject(Napi::Env env, duckdb_date_struct date_parts) {
-  auto date_parts_obj = Napi::Object::New(env);
-  date_parts_obj.Set("year", Napi::Number::New(env, date_parts.year));
-  date_parts_obj.Set("month", Napi::Number::New(env, date_parts.month));
-  date_parts_obj.Set("day", Napi::Number::New(env, date_parts.day));
-  return date_parts_obj;
-}
-
-duckdb_date_struct GetDatePartsFromObject(Napi::Object date_parts_obj) {
-  int32_t year = date_parts_obj.Get("year").As<Napi::Number>().Int32Value();
-  int8_t month = date_parts_obj.Get("month").As<Napi::Number>().Int32Value();
-  int8_t day = date_parts_obj.Get("day").As<Napi::Number>().Int32Value();
-  return { year, month, day };
-}
-
-Napi::Object MakeTimeObject(Napi::Env env, duckdb_time time) {
-  auto time_obj = Napi::Object::New(env);
-  time_obj.Set("micros", Napi::BigInt::New(env, time.micros));
-  return time_obj;
-}
-
-duckdb_time GetTimeFromObject(Napi::Env env, Napi::Object time_obj) {
-  bool lossless;
-  auto micros = time_obj.Get("micros").As<Napi::BigInt>().Int64Value(&lossless);
-  if (!lossless) {
-    throw Napi::Error::New(env, "micros out of int64 range");
-  }
-  return { micros };
-}
-
-Napi::Object MakeTimePartsObject(Napi::Env env, duckdb_time_struct time_parts) {
-  auto time_parts_obj = Napi::Object::New(env);
-  time_parts_obj.Set("hour", Napi::Number::New(env, time_parts.hour));
-  time_parts_obj.Set("min", Napi::Number::New(env, time_parts.min));
-  time_parts_obj.Set("sec", Napi::Number::New(env, time_parts.sec));
-  time_parts_obj.Set("micros", Napi::Number::New(env, time_parts.micros));
-  return time_parts_obj;
-}
-
-duckdb_time_struct GetTimePartsFromObject(Napi::Object time_parts_obj) {
-  int8_t hour = time_parts_obj.Get("hour").As<Napi::Number>().Int32Value();
-  int8_t min = time_parts_obj.Get("min").As<Napi::Number>().Int32Value();
-  int8_t sec = time_parts_obj.Get("sec").As<Napi::Number>().Int32Value();
-  int32_t micros = time_parts_obj.Get("micros").As<Napi::Number>().Int32Value();
-  return { hour, min, sec, micros };
-}
-
-Napi::Object MakeTimeNSObject(Napi::Env env, duckdb_time_ns time_ns) {
-  auto time_ns_obj = Napi::Object::New(env);
-  time_ns_obj.Set("nanos", Napi::BigInt::New(env, time_ns.nanos));
-  return time_ns_obj;
-}
-
-duckdb_time_ns GetTimeNSFromObject(Napi::Env env, Napi::Object time_ns_obj) {
-  bool lossless;
-  auto nanos = time_ns_obj.Get("nanos").As<Napi::BigInt>().Int64Value(&lossless);
-  if (!lossless) {
-    throw Napi::Error::New(env, "nanos out of int64 range");
-  }
-  return { nanos };
-}
-
-Napi::Object MakeTimeTZObject(Napi::Env env, duckdb_time_tz time_tz) {
-  auto time_tz_obj = Napi::Object::New(env);
-  time_tz_obj.Set("bits", Napi::BigInt::New(env, time_tz.bits));
-  return time_tz_obj;
-}
-
-duckdb_time_tz GetTimeTZFromObject(Napi::Env env, Napi::Object time_tz_obj) {
-  bool lossless;
-  auto bits = time_tz_obj.Get("bits").As<Napi::BigInt>().Uint64Value(&lossless);
-  if (!lossless) {
-    throw Napi::Error::New(env, "bits out of uint64 range");
-  }
-  return { bits };
-}
-
-Napi::Object MakeTimeTZPartsObject(Napi::Env env, duckdb_time_tz_struct time_tz_parts) {
-  auto time_tz_parts_obj = Napi::Object::New(env);
-  time_tz_parts_obj.Set("time", MakeTimePartsObject(env, time_tz_parts.time));
-  time_tz_parts_obj.Set("offset", Napi::Number::New(env, time_tz_parts.offset));
-  return time_tz_parts_obj;
-}
-
-// GetTimeTZPartsFromObject not used
-
-Napi::Object MakeTimestampObject(Napi::Env env, duckdb_timestamp timestamp) {
-  auto timestamp_obj = Napi::Object::New(env);
-  timestamp_obj.Set("micros", Napi::BigInt::New(env, timestamp.micros));
-  return timestamp_obj;
-}
-
-duckdb_timestamp GetTimestampFromObject(Napi::Env env, Napi::Object timestamp_obj) {
-  bool lossless;
-  auto micros = timestamp_obj.Get("micros").As<Napi::BigInt>().Int64Value(&lossless);
-  if (!lossless) {
-    throw Napi::Error::New(env, "micros out of int64 range");
-  }
-  return { micros };
-}
-
-Napi::Object MakeTimestampSecondsObject(Napi::Env env, duckdb_timestamp_s timestamp) {
-  auto timestamp_s_obj = Napi::Object::New(env);
-  timestamp_s_obj.Set("seconds", Napi::BigInt::New(env, timestamp.seconds));
-  return timestamp_s_obj;
-}
-
-duckdb_timestamp_s GetTimestampSecondsFromObject(Napi::Env env, Napi::Object timestamp_s_obj) {
-  bool lossless;
-  auto seconds = timestamp_s_obj.Get("seconds").As<Napi::BigInt>().Int64Value(&lossless);
-  if (!lossless) {
-    throw Napi::Error::New(env, "seconds out of int64 range");
-  }
-  return { seconds };
-}
-
-Napi::Object MakeTimestampMillisecondsObject(Napi::Env env, duckdb_timestamp_ms timestamp) {
-  auto timestamp_ms_obj = Napi::Object::New(env);
-  timestamp_ms_obj.Set("millis", Napi::BigInt::New(env, timestamp.millis));
-  return timestamp_ms_obj;
-}
-
-duckdb_timestamp_ms GetTimestampMillisecondsFromObject(Napi::Env env, Napi::Object timestamp_ms_obj) {
-  bool lossless;
-  auto millis = timestamp_ms_obj.Get("millis").As<Napi::BigInt>().Int64Value(&lossless);
-  if (!lossless) {
-    throw Napi::Error::New(env, "millis out of int64 range");
-  }
-  return { millis };
-}
-
-Napi::Object MakeTimestampNanosecondsObject(Napi::Env env, duckdb_timestamp_ns timestamp) {
-  auto timestamp_ns_obj = Napi::Object::New(env);
-  timestamp_ns_obj.Set("nanos", Napi::BigInt::New(env, timestamp.nanos));
-  return timestamp_ns_obj;
-}
-
-duckdb_timestamp_ns GetTimestampNanosecondsFromObject(Napi::Env env, Napi::Object timestamp_ns_obj) {
-  bool lossless;
-  auto nanos = timestamp_ns_obj.Get("nanos").As<Napi::BigInt>().Int64Value(&lossless);
-  if (!lossless) {
-    throw Napi::Error::New(env, "nanos out of int64 range");
-  }
-  return { nanos };
-}
-
-Napi::Object MakeTimestampPartsObject(Napi::Env env, duckdb_timestamp_struct timestamp_parts) {
-  auto timestamp_parts_obj = Napi::Object::New(env);
-  timestamp_parts_obj.Set("date", MakeDatePartsObject(env, timestamp_parts.date));
-  timestamp_parts_obj.Set("time", MakeTimePartsObject(env, timestamp_parts.time));
-  return timestamp_parts_obj;
-}
-
-duckdb_timestamp_struct GetTimestampPartsFromObject(Napi::Object timestamp_parts_obj) {
-  auto date = GetDatePartsFromObject(timestamp_parts_obj.Get("date").As<Napi::Object>());
-  auto time = GetTimePartsFromObject(timestamp_parts_obj.Get("time").As<Napi::Object>());
-  return { date, time };
-}
-
-Napi::Object MakeIntervalObject(Napi::Env env, duckdb_interval interval) {
-  auto interval_obj = Napi::Object::New(env);
-  interval_obj.Set("months", Napi::Number::New(env, interval.months));
-  interval_obj.Set("days", Napi::Number::New(env, interval.days));
-  interval_obj.Set("micros", Napi::BigInt::New(env, interval.micros));
-  return interval_obj;
-}
-
-duckdb_interval GetIntervalFromObject(Napi::Env env, Napi::Object interval_obj) {
-  int32_t months = interval_obj.Get("months").As<Napi::Number>().Int32Value();
-  int32_t days = interval_obj.Get("days").As<Napi::Number>().Int32Value();
-  bool lossless;
-  int64_t micros = interval_obj.Get("micros").As<Napi::BigInt>().Int64Value(&lossless);
-  if (!lossless) {
-    throw Napi::Error::New(env, "micros out of int64 range");
-  }
-  return { months, days, micros };
-}
-
-duckdb_hugeint GetHugeIntFromBigInt(Napi::Env env, Napi::BigInt bigint) {
-  int sign_bit;
-  size_t word_count = 2;
-  uint64_t words[2];
-  bigint.ToWords(&sign_bit, &word_count, words);
-  if (word_count > 2) {
-    throw Napi::Error::New(env, "bigint out of hugeint range");
-  }
-  bool carry = false;
-  uint64_t lower;
-  if (word_count > 0) {
-    lower = words[0];
-    if (sign_bit) {
-      lower = ~lower + 1;
-      carry = lower == 0;
-    }
-  } else {
-    lower = 0;
-  }
-  uint64_t upper;
-  if (word_count > 1) {
-    upper = words[1];
-    if (sign_bit) {
-      upper = ~upper;
-      if (carry) {
-        upper += 1;
-      }
-    }
-  } else {
-    if (word_count > 0 && sign_bit) {
-      upper = -1;
-    } else {
-      upper = 0;
-    }
-  }
-  return { lower, int64_t(upper) };
-}
-
-Napi::BigInt MakeBigIntFromHugeInt(Napi::Env env, duckdb_hugeint hugeint) {
-  int sign_bit = hugeint.upper < 0 ? 1 : 0;
-  size_t word_count = hugeint.upper == -1 ? 1 : 2;
-  uint64_t words[2];
-  bool carry = false;
-  words[0] = hugeint.lower;
-  if (sign_bit) {
-    words[0] = ~(words[0] - 1);
-    carry = words[0] == 0;
-  }
-  if (word_count > 1) {
-    words[1] = hugeint.upper;
-    if (sign_bit) {
-      if (carry) {
-        words[1] -= 1;
-      }
-      words[1] = ~words[1];
-    }
-  } else {
-    words[1] = 0;
-  }
-  return Napi::BigInt::New(env, sign_bit, word_count, words);
-}
-
-duckdb_uhugeint GetUHugeIntFromBigInt(Napi::Env env, Napi::BigInt bigint) {
-  int sign_bit;
-  size_t word_count = 2;
-  uint64_t words[2];
-  bigint.ToWords(&sign_bit, &word_count, words);
-  if (word_count > 2 || sign_bit) {
-    throw Napi::Error::New(env, "bigint out of uhugeint range");
-  }
-  uint64_t lower = word_count > 0 ? words[0] : 0;
-  uint64_t upper = word_count > 1 ? words[1] : 0;
-  return { lower, upper };
-}
-
-Napi::BigInt MakeBigIntFromUHugeInt(Napi::Env env, duckdb_uhugeint uhugeint) {
-  int sign_bit = 0;
-  size_t word_count = 2;
-  uint64_t words[2];
-  words[0] = uhugeint.lower;
-  words[1] = uhugeint.upper;
-  return Napi::BigInt::New(env, sign_bit, word_count, words);
-}
-
-duckdb_bignum GetBigNumFromBigInt(Napi::Env env, Napi::BigInt bigint) {
-  int sign_bit;
-  size_t word_count = bigint.WordCount();
-  size_t byte_count = word_count * 8;
-  uint64_t *words = static_cast<uint64_t*>(duckdb_malloc(byte_count));
-  bigint.ToWords(&sign_bit, &word_count, words);
-  uint8_t *data = reinterpret_cast<uint8_t*>(words);
-  idx_t size = byte_count;
-  bool is_negative = bool(sign_bit);
-  // convert little-endian to big-endian
-  for (size_t i = 0; i < size/2; i++) {
-    auto tmp = data[i];
-    data[i] = data[size - 1 - i];
-    data[size - 1 - i] = tmp;
-  }
-  return { data, size, is_negative };
-}
-
-Napi::BigInt MakeBigIntFromBigNum(Napi::Env env, duckdb_bignum bignum) {
-  int sign_bit = bignum.is_negative ? 1 : 0;
-  size_t word_count = bignum.size / 8;
-  uint8_t *data = static_cast<uint8_t*>(duckdb_malloc(bignum.size));
-  // convert big-endian to little-endian
-  for (size_t i = 0; i < bignum.size; i++) {
-    data[i] = bignum.data[bignum.size - 1 - i];
-  }
-  uint64_t *words = reinterpret_cast<uint64_t*>(data);
-  auto bigint = Napi::BigInt::New(env, sign_bit, word_count, words);
-  duckdb_free(data);
-  return bigint;
-}
-
-Napi::Object MakeDecimalObject(Napi::Env env, duckdb_decimal decimal) {
-  auto decimal_obj = Napi::Object::New(env);
-  decimal_obj.Set("width", Napi::Number::New(env, decimal.width));
-  decimal_obj.Set("scale", Napi::Number::New(env, decimal.scale));
-  decimal_obj.Set("value", MakeBigIntFromHugeInt(env, decimal.value));
-  return decimal_obj;
-}
-
-duckdb_decimal GetDecimalFromObject(Napi::Env env, Napi::Object decimal_obj) {
-  uint8_t width = decimal_obj.Get("width").As<Napi::Number>().Uint32Value();
-  uint8_t scale = decimal_obj.Get("scale").As<Napi::Number>().Uint32Value();
-  auto value = GetHugeIntFromBigInt(env, decimal_obj.Get("value").As<Napi::BigInt>());
-  return { width, scale, value };
-}
-
-// Externals
-
-template<typename T, typename Finalizer>
-Napi::External<T> CreateExternal(Napi::Env env, const napi_type_tag &type_tag, T *data, Finalizer finalizer) {
-  auto external = Napi::External<T>::New(env, data, finalizer);
-  external.TypeTag(&type_tag);
-  return external;
-}
-
-template<typename T>
-Napi::External<T> CreateExternalWithoutFinalizer(Napi::Env env, const napi_type_tag &type_tag, T *data) {
-  auto external = Napi::External<T>::New(env, data);
-  external.TypeTag(&type_tag);
-  return external;
-}
-
-template<typename T>
-T* GetDataFromExternal(Napi::Env env, const napi_type_tag &type_tag, Napi::Value value, const char *error_message) {
-  auto external = value.As<Napi::External<T>>();
-  if (!external.CheckTypeTag(&type_tag)) {
-    throw Napi::Error::New(env, error_message);
-  }
-  return external.Data();
-}
-
-// The following type tags are generated using: uuidgen | sed -r -e 's/-//g' -e 's/(.{16})(.*)/0x\1, 0x\2/'
-
-static const napi_type_tag AppenderTypeTag = {
-  0x32E0AB3B83F74A89, 0xB785905D92D54996
-};
-
-void FinalizeAppender(Napi::BasicEnv, duckdb_appender appender) {
-  if (appender) {
-    duckdb_appender_destroy(&appender);
-    appender = nullptr;
-  }
-}
-
-Napi::External<_duckdb_appender> CreateExternalForAppender(Napi::Env env, duckdb_appender appender) {
-  return CreateExternal<_duckdb_appender>(env, AppenderTypeTag, appender, FinalizeAppender);
-}
-
-duckdb_appender GetAppenderFromExternal(Napi::Env env, Napi::Value value) {
-  return GetDataFromExternal<_duckdb_appender>(env, AppenderTypeTag, value, "Invalid appender argument");
-}
-
-static const napi_type_tag BindInfoTypeTag = {
-  0x9922F8468F9A43C0, 0xB2573B112B9600D8
-};
-
-Napi::External<_duckdb_bind_info> CreateExternalForBindInfoWithoutFinalizer(Napi::Env env, duckdb_bind_info bind_info) {
-  // BindInfo objects are never explicitly created; they are passed in to function callbacks.
-  return CreateExternalWithoutFinalizer<_duckdb_bind_info>(env, BindInfoTypeTag, bind_info);
-}
-
-duckdb_bind_info GetBindInfoFromExternal(Napi::Env env, Napi::Value value) {
-  return GetDataFromExternal<_duckdb_bind_info>(env, BindInfoTypeTag, value, "Invalid bind info argument");
-}
-
-static const napi_type_tag ClientContextTypeTag = {
-  0x1E1738782ED94232, 0x867B024D1858DF3A
-};
-
-void FinalizeClientContext(Napi::BasicEnv, duckdb_client_context client_context) {
-  duckdb_destroy_client_context(&client_context);
-}
-
-Napi::External<_duckdb_client_context> CreateExternalForClientContext(Napi::Env env, duckdb_client_context client_context) {
-  return CreateExternal<_duckdb_client_context>(env, ClientContextTypeTag, client_context, FinalizeClientContext);
-}
-
-duckdb_client_context GetClientContextFromExternal(Napi::Env env, Napi::Value value) {
-  return GetDataFromExternal<_duckdb_client_context>(env, ClientContextTypeTag, value, "Invalid client context argument");
-}
-
-static const napi_type_tag ConfigTypeTag = {
-  0x5963FBB9648B4D2A, 0xB41ADE86056218D1
-};
-
-void FinalizeConfig(Napi::BasicEnv, duckdb_config config) {
-  if (config) {
-    duckdb_destroy_config(&config);
-    config = nullptr;
-  }
-}
-
-Napi::External<_duckdb_config> CreateExternalForConfig(Napi::Env env, duckdb_config config) {
-  return CreateExternal<_duckdb_config>(env, ConfigTypeTag, config, FinalizeConfig);
-}
-
-duckdb_config GetConfigFromExternal(Napi::Env env, Napi::Value value) {
-  return GetDataFromExternal<_duckdb_config>(env, ConfigTypeTag, value, "Invalid config argument");
-}
-
-static const napi_type_tag ConnectionTypeTag = {
-  0x922B9BF54AB04DFC, 0x8A258578D371DB71
-};
-
-typedef struct {
-  duckdb_connection connection;
-} duckdb_connection_holder;
-
-duckdb_connection_holder *CreateConnectionHolder(duckdb_connection connection) {
-  auto connection_holder_ptr = reinterpret_cast<duckdb_connection_holder*>(duckdb_malloc(sizeof(duckdb_connection_holder)));
-  connection_holder_ptr->connection = connection;
-  return connection_holder_ptr;
-}
-
-void FinalizeConnectionHolder(Napi::BasicEnv, duckdb_connection_holder *connection_holder_ptr) {
-  // duckdb_disconnect is a no-op if already disconnected
-  duckdb_disconnect(&connection_holder_ptr->connection);
-  duckdb_free(connection_holder_ptr);
-}
-
-Napi::External<duckdb_connection_holder> CreateExternalForConnection(Napi::Env env, duckdb_connection connection) {
-  return CreateExternal<duckdb_connection_holder>(env, ConnectionTypeTag, CreateConnectionHolder(connection), FinalizeConnectionHolder);
-}
-
-duckdb_connection_holder *GetConnectionHolderFromExternal(Napi::Env env, Napi::Value value) {
-  return GetDataFromExternal<duckdb_connection_holder>(env, ConnectionTypeTag, value, "Invalid connection argument");
-}
-
-duckdb_connection GetConnectionFromExternal(Napi::Env env, Napi::Value value) {
-  return GetConnectionHolderFromExternal(env, value)->connection;
-}
-
-static const napi_type_tag DatabaseTypeTag = {
-  0x835A8533653C40D1, 0x83B3BE2B233BA8F3
-};
-
-typedef struct {
-  duckdb_database database;
-} duckdb_database_holder;
-
-duckdb_database_holder *CreateDatabaseHolder(duckdb_database database) {
-  auto database_holder_ptr = reinterpret_cast<duckdb_database_holder*>(duckdb_malloc(sizeof(duckdb_database_holder)));
-  database_holder_ptr->database = database;
-  return database_holder_ptr;
-}
-
-void FinalizeDatabaseHolder(Napi::BasicEnv, duckdb_database_holder *database_holder_ptr) {
-  // duckdb_close is a no-op if already closed
-  duckdb_close(&database_holder_ptr->database);
-  duckdb_free(database_holder_ptr);
-}
-
-Napi::External<duckdb_database_holder> CreateExternalForDatabase(Napi::Env env, duckdb_database database) {
-  return CreateExternal<duckdb_database_holder>(env, DatabaseTypeTag, CreateDatabaseHolder(database), FinalizeDatabaseHolder);
-}
-
-duckdb_database_holder *GetDatabaseHolderFromExternal(Napi::Env env, Napi::Value value) {
-  return GetDataFromExternal<duckdb_database_holder>(env, DatabaseTypeTag, value, "Invalid database argument");
-}
-
-duckdb_database GetDatabaseFromExternal(Napi::Env env, Napi::Value value) {
-  return GetDatabaseHolderFromExternal(env, value)->database;
-}
-
-static const napi_type_tag DataChunkTypeTag = {
-  0x2C7537AB063A4296, 0xB1E70F08B0BBD1A3
-};
-
-void FinalizeDataChunk(Napi::BasicEnv, duckdb_data_chunk chunk) {
-  if (chunk) {
-    duckdb_destroy_data_chunk(&chunk);
-    chunk = nullptr;
-  }
-}
-
-Napi::External<_duckdb_data_chunk> CreateExternalForDataChunk(Napi::Env env, duckdb_data_chunk chunk) {
-  return CreateExternal<_duckdb_data_chunk>(env, DataChunkTypeTag, chunk, FinalizeDataChunk);
-}
-
-Napi::External<_duckdb_data_chunk> CreateExternalForDataChunkWithoutFinalizer(Napi::Env env, duckdb_data_chunk chunk) {
-  return CreateExternalWithoutFinalizer<_duckdb_data_chunk>(env, DataChunkTypeTag, chunk);
-}
-
-duckdb_data_chunk GetDataChunkFromExternal(Napi::Env env, Napi::Value value) {
-  return GetDataFromExternal<_duckdb_data_chunk>(env, DataChunkTypeTag, value, "Invalid data chunk argument");
-}
-
-static const napi_type_tag ExtractedStatementsTypeTag = {
-  0x59288E1C60C44EEB, 0xBFA35376EE0F04DD
-};
-
-void FinalizeExtractedStatements(Napi::BasicEnv, duckdb_extracted_statements extracted_statements) {
-  if (extracted_statements) {
-    duckdb_destroy_extracted(&extracted_statements);
-    extracted_statements = nullptr;
-  }
-}
-
-Napi::External<_duckdb_extracted_statements> CreateExternalForExtractedStatements(Napi::Env env, duckdb_extracted_statements extracted_statements) {
-  return CreateExternal<_duckdb_extracted_statements>(env, ExtractedStatementsTypeTag, extracted_statements, FinalizeExtractedStatements);
-}
-
-duckdb_extracted_statements GetExtractedStatementsFromExternal(Napi::Env env, Napi::Value value) {
-  return GetDataFromExternal<_duckdb_extracted_statements>(env, ExtractedStatementsTypeTag, value, "Invalid extracted statements argument");
-}
-
-static const napi_type_tag FunctionInfoTypeTag = {
-  0xB0E6739D698048EA, 0x9E79734E3E137AC3
-};
-
-Napi::External<_duckdb_function_info> CreateExternalForFunctionInfoWithoutFinalizer(Napi::Env env, duckdb_function_info function_info) {
-  // FunctionInfo objects are never explicitly created; they are passed in to function callbacks.
-  return CreateExternalWithoutFinalizer<_duckdb_function_info>(env, FunctionInfoTypeTag, function_info);
-}
-
-duckdb_function_info GetFunctionInfoFromExternal(Napi::Env env, Napi::Value value) {
-  return GetDataFromExternal<_duckdb_function_info>(env, FunctionInfoTypeTag, value, "Invalid function info argument");
-}
-
-static const napi_type_tag InstanceCacheTypeTag = {
-  0x2F3346E30FB5457C, 0xB9201EE5112EEF9F
-};
-
-void FinalizeInstanceCache(Napi::BasicEnv, duckdb_instance_cache instance_cache) {
-  duckdb_destroy_instance_cache(&instance_cache);
-}
-
-Napi::External<_duckdb_instance_cache> CreateExternalForInstanceCache(Napi::Env env, duckdb_instance_cache instance_cache) {
-  return CreateExternal<_duckdb_instance_cache>(env, InstanceCacheTypeTag, instance_cache, FinalizeInstanceCache);
-}
-
-duckdb_instance_cache GetInstanceCacheFromExternal(Napi::Env env, Napi::Value value) {
-  return GetDataFromExternal<_duckdb_instance_cache>(env, InstanceCacheTypeTag, value, "Invalid instance cache argument");
-}
-
-static const napi_type_tag LogicalTypeTypeTag = {
-  0x78AF202191ED4A23, 0x8093715369592A2B
-};
-
-void FinalizeLogicalType(Napi::BasicEnv, duckdb_logical_type logical_type) {
-  duckdb_destroy_logical_type(&logical_type);
-}
-
-Napi::External<_duckdb_logical_type> CreateExternalForLogicalType(Napi::Env env, duckdb_logical_type logical_type) {
-  return CreateExternal<_duckdb_logical_type>(env, LogicalTypeTypeTag, logical_type, FinalizeLogicalType);
-}
-
-Napi::External<_duckdb_logical_type> CreateExternalForLogicalTypeWithoutFinalizer(Napi::Env env, duckdb_logical_type logical_type) {
-  return CreateExternalWithoutFinalizer<_duckdb_logical_type>(env, LogicalTypeTypeTag, logical_type);
-}
-
-duckdb_logical_type GetLogicalTypeFromExternal(Napi::Env env, Napi::Value value) {
-  return GetDataFromExternal<_duckdb_logical_type>(env, LogicalTypeTypeTag, value, "Invalid logical type argument");
-}
-
-static const napi_type_tag PendingResultTypeTag = {
-  0x257E88ECE8294FEC, 0xB64963BBBD1DBB41
-};
-
-void FinalizePendingResult(Napi::BasicEnv, duckdb_pending_result pending_result) {
-  if (pending_result) {
-    duckdb_destroy_pending(&pending_result);
-    pending_result = nullptr;
-  }
-}
-
-Napi::External<_duckdb_pending_result> CreateExternalForPendingResult(Napi::Env env, duckdb_pending_result pending_result) {
-  return CreateExternal<_duckdb_pending_result>(env, PendingResultTypeTag, pending_result, FinalizePendingResult);
-}
-
-duckdb_pending_result GetPendingResultFromExternal(Napi::Env env, Napi::Value value) {
-  return GetDataFromExternal<_duckdb_pending_result>(env, PendingResultTypeTag, value, "Invalid pending result argument");
-}
-
-static const napi_type_tag PreparedStatementTypeTag = {
-  0xA8B03DAD16D34416, 0x9735A7E1F2A1240C
-};
-
-typedef struct {
-  duckdb_prepared_statement prepared;
-} duckdb_prepared_statement_holder;
-
-duckdb_prepared_statement_holder *CreatePreparedStatementHolder(duckdb_prepared_statement prepared) {
-  auto prepared_statement_holder_ptr = reinterpret_cast<duckdb_prepared_statement_holder*>(duckdb_malloc(sizeof(duckdb_prepared_statement_holder)));
-  prepared_statement_holder_ptr->prepared = prepared;
-  return prepared_statement_holder_ptr;
-}
-
-void FinalizePreparedStatementHolder(Napi::BasicEnv, duckdb_prepared_statement_holder *prepared_statement_holder_ptr) {
-  // duckdb_destroy_prepare is a no-op if already destroyed
-  duckdb_destroy_prepare(&prepared_statement_holder_ptr->prepared);
-  duckdb_free(prepared_statement_holder_ptr);
-}
-
-Napi::External<duckdb_prepared_statement_holder> CreateExternalForPreparedStatement(Napi::Env env, duckdb_prepared_statement prepared_statement) {
-  return CreateExternal<duckdb_prepared_statement_holder>(env, PreparedStatementTypeTag, CreatePreparedStatementHolder(prepared_statement), FinalizePreparedStatementHolder);
-}
-
-duckdb_prepared_statement_holder *GetPreparedStatementHolderFromExternal(Napi::Env env, Napi::Value value) {
-  return GetDataFromExternal<duckdb_prepared_statement_holder>(env, PreparedStatementTypeTag, value, "Invalid prepared statement argument");
-}
-
-duckdb_prepared_statement GetPreparedStatementFromExternal(Napi::Env env, Napi::Value value) {
-  return GetPreparedStatementHolderFromExternal(env, value)->prepared;
-}
-
-static const napi_type_tag ResultTypeTag = {
-  0x08F7FE3AE12345E5, 0x8733310DC29372D9
-};
-
-void FinalizeResult(Napi::BasicEnv, duckdb_result *result_ptr) {
-  if (result_ptr) {
-    duckdb_destroy_result(result_ptr);
-    duckdb_free(result_ptr); // memory for duckdb_result struct is malloc'd in QueryWorker, ExecutePreparedWorker, or ExecutePendingWorker.
-    result_ptr = nullptr;
-  }
-}
-
-Napi::External<duckdb_result> CreateExternalForResult(Napi::Env env, duckdb_result *result_ptr) {
-  return CreateExternal<duckdb_result>(env, ResultTypeTag, result_ptr, FinalizeResult);
-}
-
-duckdb_result *GetResultFromExternal(Napi::Env env, Napi::Value value) {
-  return GetDataFromExternal<duckdb_result>(env, ResultTypeTag, value, "Invalid result argument");
-}
-
-static const napi_type_tag ScalarFunctionTypeTag = {
-  0x95D48B7051D14994, 0x9F883D7DF5DEA86D
-};
-
-using ScalarFunctionBindTSFNContext = std::nullptr_t;
-struct ScalarFunctionBindTSFNData;
-void ScalarFunctionBindTSFNCallback(Napi::Env env, Napi::Function callback, ScalarFunctionBindTSFNContext *context, ScalarFunctionBindTSFNData *data);
-using ScalarFunctionBindTSFN = Napi::TypedThreadSafeFunction<ScalarFunctionBindTSFNContext, ScalarFunctionBindTSFNData, ScalarFunctionBindTSFNCallback>;
-
-using ScalarFunctionMainTSFNContext = std::nullptr_t;
-struct ScalarFunctionMainTSFNData;
-void ScalarFunctionMainTSFNCallback(Napi::Env env, Napi::Function callback, ScalarFunctionMainTSFNContext *context, ScalarFunctionMainTSFNData *data);
-using ScalarFunctionMainTSFN = Napi::TypedThreadSafeFunction<ScalarFunctionMainTSFNContext, ScalarFunctionMainTSFNData, ScalarFunctionMainTSFNCallback>;
-
-struct ScalarFunctionInternalExtraInfo {
-  std::unique_ptr<ScalarFunctionBindTSFN> bind_tsfn;
-  std::unique_ptr<ScalarFunctionMainTSFN> main_tsfn;
-  std::unique_ptr<Napi::ObjectReference> user_extra_info_ref;
-
-  ScalarFunctionInternalExtraInfo() {}
-
-  ~ScalarFunctionInternalExtraInfo() {
-    if (bool(bind_tsfn)) {
-      bind_tsfn->Release();
-    }
-    if (bool(main_tsfn)) {
-      main_tsfn->Release();
-    }
-  }
-
-  void SetBindFunction(Napi::Env env, Napi::Function func) {
-    if (bool(bind_tsfn)) {
-      bind_tsfn->Release();
-    }
-    bind_tsfn = std::make_unique<ScalarFunctionBindTSFN>(ScalarFunctionBindTSFN::New(env, func, "ScalarFunctionBind", 0, 1));
-  }
-
-  void SetMainFunction(Napi::Env env, Napi::Function func) {
-    if (bool(main_tsfn)) {
-      main_tsfn->Release();
-    }
-    main_tsfn = std::make_unique<ScalarFunctionMainTSFN>(ScalarFunctionMainTSFN::New(env, func, "ScalarFunctionMain", 0, 1));
-  }
-
-  void SetUserExtraInfo(Napi::Object user_extra_info) {
-    user_extra_info_ref = std::make_unique<Napi::ObjectReference>(user_extra_info.IsUndefined() ? Napi::ObjectReference() : Napi::Persistent(user_extra_info));
-  }
-};
-
-void DeleteScalarFunctionInternalExtraInfo(ScalarFunctionInternalExtraInfo *internal_extra_info) {
-  delete internal_extra_info;
-}
-
-struct ScalarFunctionHolder {
-  duckdb_scalar_function scalar_function;
-  ScalarFunctionInternalExtraInfo *internal_extra_info;
-
-  ScalarFunctionHolder(duckdb_scalar_function scalar_function_in): scalar_function(scalar_function_in), internal_extra_info(nullptr) {}
-
-  ~ScalarFunctionHolder() {
-    // duckdb_destroy_scalar_function is a no-op if already destroyed
-    duckdb_destroy_scalar_function(&scalar_function);
-  }
-
-  ScalarFunctionInternalExtraInfo *EnsureInternalExtraInfo() {
-    if (!internal_extra_info) {
-      internal_extra_info = new ScalarFunctionInternalExtraInfo();
-      duckdb_scalar_function_set_extra_info(scalar_function, internal_extra_info, reinterpret_cast<duckdb_delete_callback_t>(DeleteScalarFunctionInternalExtraInfo));
-    }
-    return internal_extra_info;
-  }
-};
-
-ScalarFunctionHolder *CreateScalarFunctionHolder(duckdb_scalar_function scalar_function) {
-  return new ScalarFunctionHolder(scalar_function);
-}
-
-void FinalizeScalarFunctionHolder(Napi::BasicEnv, ScalarFunctionHolder *holder) {
-  delete holder;
-}
-
-Napi::External<ScalarFunctionHolder> CreateExternalForScalarFunction(Napi::Env env, duckdb_scalar_function scalar_function) {
-  return CreateExternal<ScalarFunctionHolder>(env, ScalarFunctionTypeTag, CreateScalarFunctionHolder(scalar_function), FinalizeScalarFunctionHolder);
-}
-
-ScalarFunctionHolder *GetScalarFunctionHolderFromExternal(Napi::Env env, Napi::Value value) {
-  return GetDataFromExternal<ScalarFunctionHolder>(env, ScalarFunctionTypeTag, value, "Invalid scalar function argument");
-}
-
-duckdb_scalar_function GetScalarFunctionFromExternal(Napi::Env env, Napi::Value value) {
-  return GetScalarFunctionHolderFromExternal(env, value)->scalar_function;
-}
-
-static const napi_type_tag ValueTypeTag = {
-  0xC60F36613BF14E93, 0xBAA92848936FAA25
-};
-
-void FinalizeValue(Napi::BasicEnv, duckdb_value value) {
-  if (value) {
-    duckdb_destroy_value(&value);
-    value = nullptr;
-  }
-}
-
-Napi::External<_duckdb_value> CreateExternalForValue(Napi::Env env, duckdb_value value) {
-  return CreateExternal<_duckdb_value>(env, ValueTypeTag, value, FinalizeValue);
-}
-
-duckdb_value GetValueFromExternal(Napi::Env env, Napi::Value value) {
-  return GetDataFromExternal<_duckdb_value>(env, ValueTypeTag, value, "Invalid value argument");
-}
-
-static const napi_type_tag VectorTypeTag = {
-  0x9FE56DE8E3124D07, 0x9ABF31145EDE1C9E
-};
-
-Napi::External<_duckdb_vector> CreateExternalForVectorWithoutFinalizer(Napi::Env env, duckdb_vector vector) {
-  // Vectors live as long as their containing data chunk; they cannot be explicitly destroyed.
-  return CreateExternalWithoutFinalizer<_duckdb_vector>(env, VectorTypeTag, vector);
-}
-
-duckdb_vector GetVectorFromExternal(Napi::Env env, Napi::Value value) {
-  return GetDataFromExternal<_duckdb_vector>(env, VectorTypeTag, value, "Invalid vector argument");
-}
-
-// Scalar functions
-
-struct ScalarFunctionInternalBindData {
-  std::unique_ptr<Napi::ObjectReference> user_bind_data_ref;
-
-  void SetUserBindData(Napi::Object user_bind_data) {
-    user_bind_data_ref = std::make_unique<Napi::ObjectReference>(user_bind_data.IsUndefined() ? Napi::ObjectReference() : Napi::Persistent(user_bind_data));
-  }
-};
-
-void DeleteScalarFunctionInternalBindData(ScalarFunctionInternalBindData *internal_bind_data) {
-  delete internal_bind_data;
-}
-
-ScalarFunctionInternalBindData *CopyScalarFunctionInternalBindData(ScalarFunctionInternalBindData *internal_bind_data) {
-  if (!internal_bind_data) {
-    return nullptr;
-  }
-  auto new_internal_bind_data = new ScalarFunctionInternalBindData();
-  if (internal_bind_data->user_bind_data_ref && !internal_bind_data->user_bind_data_ref->IsEmpty()) {
-    new_internal_bind_data->SetUserBindData(internal_bind_data->user_bind_data_ref->Value());
-  }
-  return new_internal_bind_data;
-}
-
-struct ScalarFunctionBindTSFNData {
-  duckdb_bind_info info;
-  std::condition_variable *cv;
-  std::mutex *cv_mutex;
-  bool done;
-};
-
-void ScalarFunctionBindTSFNCallback(Napi::Env env, Napi::Function callback, ScalarFunctionBindTSFNContext *context, ScalarFunctionBindTSFNData *data) {
-  if (env != nullptr) {
-    if (callback != nullptr) {
-      try {
-        callback.Call(
-          env.Undefined(),
-          {
-            CreateExternalForBindInfoWithoutFinalizer(env, data->info)
-          }
-        );
-      } catch (const Napi::Error &err) {
-        duckdb_scalar_function_bind_set_error(data->info, err.Message().c_str());
-      }
-    }
-  }
-  {
-    std::lock_guard lk(*data->cv_mutex);
-    data->done = true;
-    data->cv->notify_one();
-  } 
-}
-
-ScalarFunctionInternalExtraInfo *GetScalarFunctionInternalExtraInfoFromBindInfo(duckdb_bind_info bind_info) {
-  return reinterpret_cast<ScalarFunctionInternalExtraInfo*>(duckdb_scalar_function_bind_get_extra_info(bind_info));
-}
-
-void ScalarFunctionBindFunction(duckdb_bind_info info) {
-  auto internal_extra_info = GetScalarFunctionInternalExtraInfoFromBindInfo(info);
-  auto data = reinterpret_cast<ScalarFunctionBindTSFNData*>(duckdb_malloc(sizeof(ScalarFunctionBindTSFNData)));
-  data->info = info;
-  data->cv = new std::condition_variable;
-  data->cv_mutex = new std::mutex;
-  data->done = false;
-  // The "blocking" part of this call only waits for queue space, not for the JS function call to complete.
-  // Since we specify no limit to the queue space, it in fact never blocks.
-  auto status = internal_extra_info->bind_tsfn->BlockingCall(data);
-  if (status == napi_ok) {
-    // Wait for the JS function call to complete.
-    std::unique_lock<std::mutex> lk(*data->cv_mutex);
-    data->cv->wait(lk, [&]{ return data->done; });
-  } else {
-    duckdb_scalar_function_bind_set_error(info, "BlockingCall returned not ok");
-  }
-  delete data->cv;
-  delete data->cv_mutex;
-  duckdb_free(data);
-}
-
-struct ScalarFunctionMainTSFNData {
-  duckdb_function_info info;
-  duckdb_data_chunk input;
-  duckdb_vector output;
-  std::condition_variable *cv;
-  std::mutex *cv_mutex;
-  bool done;
-};
-
-void ScalarFunctionMainTSFNCallback(Napi::Env env, Napi::Function callback, ScalarFunctionMainTSFNContext *context, ScalarFunctionMainTSFNData *data) {
-  if (env != nullptr) {
-    if (callback != nullptr) {
-      try {
-        callback.Call(
-          env.Undefined(),
-          {
-            CreateExternalForFunctionInfoWithoutFinalizer(env, data->info),
-            CreateExternalForDataChunkWithoutFinalizer(env, data->input),
-            CreateExternalForVectorWithoutFinalizer(env, data->output)
-          }
-        );
-      } catch (const Napi::Error &err) {
-        duckdb_scalar_function_set_error(data->info, err.Message().c_str());
-      }
-    }
-  }
-  {
-    std::lock_guard lk(*data->cv_mutex);
-    data->done = true;
-    data->cv->notify_one();
-  }
-}
-
-ScalarFunctionInternalExtraInfo *GetScalarFunctionInternalExtraInfoFromFunctionInfo(duckdb_function_info function_info) {
-  return reinterpret_cast<ScalarFunctionInternalExtraInfo*>(duckdb_scalar_function_get_extra_info(function_info));
-}
-
-void ScalarFunctionMainFunction(duckdb_function_info info, duckdb_data_chunk input, duckdb_vector output) {
-  auto internal_extra_info = GetScalarFunctionInternalExtraInfoFromFunctionInfo(info);
-  auto data = reinterpret_cast<ScalarFunctionMainTSFNData*>(duckdb_malloc(sizeof(ScalarFunctionMainTSFNData)));
-  data->info = info;
-  data->input = input;
-  data->output = output;
-  data->cv = new std::condition_variable;
-  data->cv_mutex = new std::mutex;
-  data->done = false;
-  // The "blocking" part of this call only waits for queue space, not for the JS function call to complete.
-  // Since we specify no limit to the queue space, it in fact never blocks.
-  auto status = internal_extra_info->main_tsfn->BlockingCall(data);
-  if (status == napi_ok) {
-    // Wait for the JS function call to complete.
-    std::unique_lock<std::mutex> lk(*data->cv_mutex);
-    data->cv->wait(lk, [&]{ return data->done; });
-  } else {
-    duckdb_scalar_function_set_error(info, "BlockingCall returned not ok");
-  }
-  delete data->cv;
-  delete data->cv_mutex;
-  duckdb_free(data);
-}
-
-// Promise workers
-
-class PromiseWorker : public Napi::AsyncWorker {
-
-public:
-
-  PromiseWorker(Napi::Env env) : AsyncWorker(env), deferred_(Napi::Promise::Deferred::New(env)) {
-  }
-
-  Napi::Promise Promise() {
-    return deferred_.Promise();
-  }
-
-protected:
-
-  virtual Napi::Value Result() = 0;
-
-  void OnOK() override {
-		deferred_.Resolve(Result());
-	}
-
-	void OnError(const Napi::Error &e) override {
-		deferred_.Reject(e.Value());
-	}
-
-private:
-
-	Napi::Promise::Deferred deferred_;
-
-};
-
-class GetOrCreateFromCacheWorker : public PromiseWorker {
-
-public:
-
-  GetOrCreateFromCacheWorker(Napi::Env env, Napi::Value instanceCacheValue, std::optional<std::string> path, Napi::Value configValue)
-      : PromiseWorker(env),
-      instance_cache_(GetInstanceCacheFromExternal(env, instanceCacheValue)),
-      instanceCacheValueRef_(MakeValueRef(instanceCacheValue)),
-      path_(path),
-      config_(configValue.IsUndefined() ? nullptr : GetConfigFromExternal(env, configValue)),
-      configValueRef_(MakeValueRef(configValue))
-    {
-    }
-
-protected:
-
-  void Execute() override {
-    const char *path = nullptr;
-    if (path_) {
-      path = path_->c_str();
-    }
-    duckdb_config cfg = config_;
-    // If no config was provided, create one with the default duckdb_api value.
-    if (cfg == nullptr) {
-      if (duckdb_create_config(&cfg)) {
-        duckdb_destroy_config(&cfg);
-        SetError("Failed to create config");
-        return;
-      }
-      if (duckdb_set_config(cfg, "duckdb_api", DEFAULT_DUCKDB_API)) {
-        SetError("Failed to set duckdb_api");
-        return;
-      }
-    }
-    char *error = nullptr;
-    if (duckdb_get_or_create_from_cache(instance_cache_, path, &database_, cfg, &error)) {
-      if (error != nullptr) {
-        SetError(error);
-        duckdb_free(error);
-      } else {
-        SetError("Failed to get or create from cache");
-      }
-    }
-  }
-
-  Napi::Value Result() override {
-    return CreateExternalForDatabase(Env(), database_);
-  }
-
-private:
-
-  duckdb_instance_cache instance_cache_;
-  Napi::Reference<Napi::Value> instanceCacheValueRef_;
-  std::optional<std::string> path_;
-  duckdb_config config_;
-  Napi::Reference<Napi::Value> configValueRef_;
-  duckdb_database database_ = nullptr;
-
-};
-
-class OpenWorker : public PromiseWorker {
-
-public:
-
-  OpenWorker(Napi::Env env, std::optional<std::string> path, Napi::Value configValue)
-    : PromiseWorker(env),
-    path_(path),
-    config_(configValue.IsUndefined() ? nullptr : GetConfigFromExternal(env, configValue)),
-    configValueRef_(MakeValueRef(configValue))
-  {
-  }
-
-protected:
-
-  void Execute() override {
-    const char *path = nullptr;
-    if (path_) {
-      path = path_->c_str();
-    }
-    duckdb_config cfg = config_;
-    // If no config was provided, create one with the default duckdb_api value.
-    if (cfg == nullptr) {
-      if (duckdb_create_config(&cfg)) {
-        duckdb_destroy_config(&cfg);
-        SetError("Failed to create config");
-        return;
-      }
-      if (duckdb_set_config(cfg, "duckdb_api", DEFAULT_DUCKDB_API)) {
-        SetError("Failed to set duckdb_api");
-        return;
-      }
-    }
-    char *error = nullptr;
-    if (duckdb_open_ext(path, &database_, cfg, &error)) {
-      if (error != nullptr) {
-        SetError(error);
-        duckdb_free(error);
-      } else {
-        SetError("Failed to open");
-      }
-    }
-  }
-
-  Napi::Value Result() override {
-    return CreateExternalForDatabase(Env(), database_);
-  }
-
-private:
-
-  std::optional<std::string> path_;
-  duckdb_config config_;
-  Napi::Reference<Napi::Value> configValueRef_;
-  duckdb_database database_ = nullptr;
-
-};
-
-class ConnectWorker : public PromiseWorker {
-
-public:
-
-  ConnectWorker(Napi::Env env, Napi::Value databaseValue)
-    : PromiseWorker(env),
-    database_(GetDatabaseFromExternal(env, databaseValue)),
-    databaseValueRef_(MakeValueRef(databaseValue))
-  {
-  }
-
-protected:
-
-  void Execute() override {
-    if (!database_) {
-      SetError("Failed to connect: instance closed");
-      return;
-    }
-    if (duckdb_connect(database_, &connection_)) {
-      SetError("Failed to connect");
-    }
-  }
-
-  Napi::Value Result() override {
-    return CreateExternalForConnection(Env(), connection_);
-  }
-
-private:
-
-  duckdb_database database_;
-  Napi::Reference<Napi::Value> databaseValueRef_;
-  duckdb_connection connection_ = nullptr;
-
-};
-
-class QueryWorker : public PromiseWorker {
-
-public:
-
-  QueryWorker(Napi::Env env, Napi::Value connectionValue, std::string query)
-    : PromiseWorker(env),
-    connection_(GetConnectionFromExternal(env, connectionValue)),
-    connectionValueRef_(MakeValueRef(connectionValue)),
-    query_(query)
-  {
-  }
-
-protected:
-
-  void Execute() override {
-    if (!connection_) {
-      SetError("Failed to query: connection disconnected");
-      return;
-    }
-    result_ptr_ = reinterpret_cast<duckdb_result*>(duckdb_malloc(sizeof(duckdb_result)));
-    result_ptr_->internal_data = nullptr;
-    result_ptr_->deprecated_columns = nullptr;
-    if (duckdb_query(connection_, query_.c_str(), result_ptr_)) {
-      auto error = duckdb_result_error(result_ptr_);
-      if (error) {
-        SetError(error);
-      } else {
-        SetError("Failed to query");
-      }
-      duckdb_destroy_result(result_ptr_);
-      duckdb_free(result_ptr_);
-      result_ptr_ = nullptr;
-    }
-  }
-
-  Napi::Value Result() override {
-    return CreateExternalForResult(Env(), result_ptr_);
-  }
-
-private:
-
-  duckdb_connection connection_;
-  Napi::Reference<Napi::Value> connectionValueRef_;
-  std::string query_;
-  duckdb_result *result_ptr_ = nullptr;
-
-};
-
-class PrepareWorker : public PromiseWorker {
-
-public:
-
-  PrepareWorker(Napi::Env env, Napi::Value connectionValue, std::string query)
-    : PromiseWorker(env),
-    connection_(GetConnectionFromExternal(env, connectionValue)),
-    connectionValueRef_(MakeValueRef(connectionValue)),
-    query_(query)
-  {
-  }
-
-protected:
-
-  void Execute() override {
-    if (!connection_) {
-      SetError("Failed to prepare: connection disconnected");
-      return;
-    }
-    if (duckdb_prepare(connection_, query_.c_str(), &prepared_statement_)) {
-      if (prepared_statement_) {
-        SetError(duckdb_prepare_error(prepared_statement_));
-        duckdb_destroy_prepare(&prepared_statement_);
-      } else {
-        SetError("Failed to prepare");
-      }
-    }
-  }
-
-  Napi::Value Result() override {
-    return CreateExternalForPreparedStatement(Env(), prepared_statement_);
-  }
-
-private:
-
-  duckdb_connection connection_;
-  Napi::Reference<Napi::Value> connectionValueRef_;
-  std::string query_;
-  duckdb_prepared_statement prepared_statement_ = nullptr;
-
-};
-
-class ExecutePreparedWorker : public PromiseWorker {
-
-public:
-
-  ExecutePreparedWorker(Napi::Env env, Napi::Value preparedStatementValue)
-    : PromiseWorker(env),
-    prepared_statement_(GetPreparedStatementFromExternal(env, preparedStatementValue)),
-    preparedStatementValueRef_(MakeValueRef(preparedStatementValue))
-  {
-  }
-
-protected:
-
-  void Execute() override {
-    result_ptr_ = reinterpret_cast<duckdb_result*>(duckdb_malloc(sizeof(duckdb_result)));
-    result_ptr_->internal_data = nullptr;
-    result_ptr_->deprecated_columns = nullptr;
-    if (duckdb_execute_prepared(prepared_statement_, result_ptr_)) {
-      auto error = duckdb_result_error(result_ptr_);
-      if (error) {
-        SetError(error);
-      } else {
-        SetError("Failed to execute prepared statement");
-      }
-      duckdb_destroy_result(result_ptr_);
-      duckdb_free(result_ptr_);
-      result_ptr_ = nullptr;
-    }
-  }
-
-  Napi::Value Result() override {
-    return CreateExternalForResult(Env(), result_ptr_);
-  }
-
-private:
-
-  duckdb_prepared_statement prepared_statement_;
-  Napi::Reference<Napi::Value> preparedStatementValueRef_;
-  duckdb_result *result_ptr_ = nullptr;
-
-};
-
-class ExecutePreparedStreamingWorker : public PromiseWorker {
-
-public:
-
-  ExecutePreparedStreamingWorker(Napi::Env env, Napi::Value preparedStatementValue)
-    : PromiseWorker(env),
-    prepared_statement_(GetPreparedStatementFromExternal(env, preparedStatementValue)),
-    preparedStatementValueRef_(MakeValueRef(preparedStatementValue))
-  {
-  }
-
-protected:
-
-  void Execute() override {
-    result_ptr_ = reinterpret_cast<duckdb_result*>(duckdb_malloc(sizeof(duckdb_result)));
-    result_ptr_->internal_data = nullptr;
-    result_ptr_->deprecated_columns = nullptr;
-    if (duckdb_execute_prepared_streaming(prepared_statement_, result_ptr_)) {
-      auto error = duckdb_result_error(result_ptr_);
-      if (error) {
-        SetError(error);
-      } else {
-        SetError("Failed to execute prepared statement");
-      }
-      duckdb_destroy_result(result_ptr_);
-      duckdb_free(result_ptr_);
-      result_ptr_ = nullptr;
-    }
-  }
-
-  Napi::Value Result() override {
-    return CreateExternalForResult(Env(), result_ptr_);
-  }
-
-private:
-
-  duckdb_prepared_statement prepared_statement_;
-  Napi::Reference<Napi::Value> preparedStatementValueRef_;
-  duckdb_result *result_ptr_;
-
-};
-
-class ExtractStatementsWorker : public PromiseWorker {
-
-public:
-
-  ExtractStatementsWorker(Napi::Env env, Napi::Value connectionValue, std::string query)
-    : PromiseWorker(env),
-    connection_(GetConnectionFromExternal(env, connectionValue)),
-    connectionValueRef_(MakeValueRef(connectionValue)),
-    query_(query)
-  {
-  }
-
-protected:
-
-  void Execute() override {
-    if (!connection_) {
-      SetError("Failed to extract statements: connection disconnected");
-      return;
-    }
-    statement_count_ = duckdb_extract_statements(connection_, query_.c_str(), &extracted_statements_);
-  }
-
-  Napi::Value Result() override {
-    auto extracted_statements_and_count_obj = Napi::Object::New(Env());
-    extracted_statements_and_count_obj.Set("extracted_statements", CreateExternalForExtractedStatements(Env(), extracted_statements_));
-    extracted_statements_and_count_obj.Set("statement_count", Napi::Number::New(Env(), statement_count_));
-    return extracted_statements_and_count_obj;
-  }
-
-private:
-
-  duckdb_connection connection_;
-  Napi::Reference<Napi::Value> connectionValueRef_;
-  std::string query_;
-  duckdb_extracted_statements extracted_statements_ = nullptr;
-  idx_t statement_count_ = 0;
-
-};
-
-class PrepareExtractedStatementWorker : public PromiseWorker {
-
-public:
-
-  PrepareExtractedStatementWorker(Napi::Env env, Napi::Value connectionValue, Napi::Value extractedStatementsValue, idx_t index)
-    : PromiseWorker(env),
-    connection_(GetConnectionFromExternal(env, connectionValue)),
-    connectionValueRef_(MakeValueRef(connectionValue)),
-    extracted_statements_(GetExtractedStatementsFromExternal(env, extractedStatementsValue)),
-    extractedStatementsValueRef_(MakeValueRef(extractedStatementsValue)),
-    index_(index)
-  {
-  }
-
-protected:
-
-  void Execute() override {
-    if (!connection_) {
-      SetError("Failed to prepare extracted statement: connection disconnected");
-      return;
-    }
-    if (duckdb_prepare_extracted_statement(connection_, extracted_statements_, index_, &prepared_statement_)) {
-      if (prepared_statement_) {
-        SetError(duckdb_prepare_error(prepared_statement_));
-        duckdb_destroy_prepare(&prepared_statement_);
-      } else {
-        SetError("Failed to prepare extracted statement");
-      }
-    }
-  }
-
-  Napi::Value Result() override {
-    return CreateExternalForPreparedStatement(Env(), prepared_statement_);
-  }
-
-private:
-
-  duckdb_connection connection_;
-  Napi::Reference<Napi::Value> connectionValueRef_;
-  duckdb_extracted_statements extracted_statements_;
-  Napi::Reference<Napi::Value> extractedStatementsValueRef_;
-  idx_t index_;
-  duckdb_prepared_statement prepared_statement_ = nullptr;
-
-};
-
-class ExecutePendingWorker : public PromiseWorker {
-
-public:
-
-  ExecutePendingWorker(Napi::Env env, Napi::Value pendingResultValue)
-    : PromiseWorker(env),
-    pending_result_(GetPendingResultFromExternal(env, pendingResultValue)),
-    pendingResultValueRef_(MakeValueRef(pendingResultValue))
-  {
-  }
-
-protected:
-
-  void Execute() override {
-    result_ptr_ = reinterpret_cast<duckdb_result*>(duckdb_malloc(sizeof(duckdb_result)));
-    result_ptr_->internal_data = nullptr;
-    result_ptr_->deprecated_columns = nullptr;
-    if (duckdb_execute_pending(pending_result_, result_ptr_)) {
-      auto error = duckdb_result_error(result_ptr_);
-      if (error) {
-        SetError(error);
-      } else {
-        SetError("Failed to execute pending result");
-      }
-      duckdb_destroy_result(result_ptr_);
-      duckdb_free(result_ptr_);
-      result_ptr_ = nullptr;
-    }
-  }
-
-  Napi::Value Result() override {
-    return CreateExternalForResult(Env(), result_ptr_);
-  }
-
-private:
-
-  duckdb_pending_result pending_result_;
-  Napi::Reference<Napi::Value> pendingResultValueRef_;
-  duckdb_result *result_ptr_ = nullptr;
-
-};
-
-class FetchWorker : public PromiseWorker {
-
-public:
-
-  FetchWorker(Napi::Env env, Napi::Value resultValue)
-    : PromiseWorker(env),
-    result_ptr_(GetResultFromExternal(env, resultValue)),
-    resultValueRef_(MakeValueRef(resultValue))
-  {
-  }
-
-protected:
-
-  void Execute() override {
-    data_chunk_ = duckdb_fetch_chunk(*result_ptr_);
-  }
-
-  Napi::Value Result() override {
-    return CreateExternalForDataChunk(Env(), data_chunk_);
-  }
-
-private:
-
-  duckdb_result *result_ptr_;
-  Napi::Reference<Napi::Value> resultValueRef_;
-  duckdb_data_chunk data_chunk_ = nullptr;
-
-};
-
-// Enums
-
-void DefineEnumMember(Napi::Object enumObj, const char *key, uint32_t value) {
-  enumObj.Set(key, value);
-  enumObj.Set(value, key);
-}
-
-Napi::Object CreatePendingStateEnum(Napi::Env env) {
-  auto pendingStateEnum = Napi::Object::New(env);
-  DefineEnumMember(pendingStateEnum, "RESULT_READY", 0);
-  DefineEnumMember(pendingStateEnum, "RESULT_NOT_READY", 1);
-  DefineEnumMember(pendingStateEnum, "ERROR", 2);
-  DefineEnumMember(pendingStateEnum, "NO_TASKS_AVAILABLE", 3);
-  return pendingStateEnum;
-}
-
-Napi::Object CreateResultTypeEnum(Napi::Env env) {
-  auto resultTypeEnum = Napi::Object::New(env);
-  DefineEnumMember(resultTypeEnum, "INVALID", 0);
-  DefineEnumMember(resultTypeEnum, "CHANGED_ROWS", 1);
-  DefineEnumMember(resultTypeEnum, "NOTHING", 2);
-  DefineEnumMember(resultTypeEnum, "QUERY_RESULT", 3);
-  return resultTypeEnum;
-}
-
-Napi::Object CreateStatementTypeEnum(Napi::Env env) {
-  auto statementTypeEnum = Napi::Object::New(env);
-  DefineEnumMember(statementTypeEnum, "INVALID", 0);
-	DefineEnumMember(statementTypeEnum, "SELECT", 1);
-	DefineEnumMember(statementTypeEnum, "INSERT", 2);
-	DefineEnumMember(statementTypeEnum, "UPDATE", 3);
-	DefineEnumMember(statementTypeEnum, "EXPLAIN", 4);
-	DefineEnumMember(statementTypeEnum, "DELETE", 5);
-	DefineEnumMember(statementTypeEnum, "PREPARE", 6);
-	DefineEnumMember(statementTypeEnum, "CREATE", 7);
-	DefineEnumMember(statementTypeEnum, "EXECUTE", 8);
-	DefineEnumMember(statementTypeEnum, "ALTER", 9);
-	DefineEnumMember(statementTypeEnum, "TRANSACTION", 10);
-	DefineEnumMember(statementTypeEnum, "COPY", 11);
-	DefineEnumMember(statementTypeEnum, "ANALYZE", 12);
-	DefineEnumMember(statementTypeEnum, "VARIABLE_SET", 13);
-	DefineEnumMember(statementTypeEnum, "CREATE_FUNC", 14);
-	DefineEnumMember(statementTypeEnum, "DROP", 15);
-	DefineEnumMember(statementTypeEnum, "EXPORT", 16);
-	DefineEnumMember(statementTypeEnum, "PRAGMA", 17);
-	DefineEnumMember(statementTypeEnum, "VACUUM", 18);
-	DefineEnumMember(statementTypeEnum, "CALL", 19);
-	DefineEnumMember(statementTypeEnum, "SET", 20);
-	DefineEnumMember(statementTypeEnum, "LOAD", 21);
-	DefineEnumMember(statementTypeEnum, "RELATION", 22);
-	DefineEnumMember(statementTypeEnum, "EXTENSION", 23);
-	DefineEnumMember(statementTypeEnum, "LOGICAL_PLAN", 24);
-	DefineEnumMember(statementTypeEnum, "ATTACH", 25);
-	DefineEnumMember(statementTypeEnum, "DETACH", 26);
-	DefineEnumMember(statementTypeEnum, "MULTI", 27);
-  return statementTypeEnum;
-}
-
-Napi::Object CreateTypeEnum(Napi::Env env) {
-  auto typeEnum = Napi::Object::New(env);
-  DefineEnumMember(typeEnum, "INVALID", 0);
-	DefineEnumMember(typeEnum, "BOOLEAN", 1);
-	DefineEnumMember(typeEnum, "TINYINT", 2);
-	DefineEnumMember(typeEnum, "SMALLINT", 3);
-	DefineEnumMember(typeEnum, "INTEGER", 4);
-	DefineEnumMember(typeEnum, "BIGINT", 5);
-	DefineEnumMember(typeEnum, "UTINYINT", 6);
-	DefineEnumMember(typeEnum, "USMALLINT", 7);
-	DefineEnumMember(typeEnum, "UINTEGER", 8);
-	DefineEnumMember(typeEnum, "UBIGINT", 9);
-	DefineEnumMember(typeEnum, "FLOAT", 10);
-	DefineEnumMember(typeEnum, "DOUBLE", 11);
-	DefineEnumMember(typeEnum, "TIMESTAMP", 12);
-	DefineEnumMember(typeEnum, "DATE", 13);
-	DefineEnumMember(typeEnum, "TIME", 14);
-	DefineEnumMember(typeEnum, "INTERVAL", 15);
-	DefineEnumMember(typeEnum, "HUGEINT", 16);
-	DefineEnumMember(typeEnum, "UHUGEINT", 32);
-	DefineEnumMember(typeEnum, "VARCHAR", 17);
-	DefineEnumMember(typeEnum, "BLOB", 18);
-	DefineEnumMember(typeEnum, "DECIMAL", 19);
-	DefineEnumMember(typeEnum, "TIMESTAMP_S", 20);
-	DefineEnumMember(typeEnum, "TIMESTAMP_MS", 21);
-	DefineEnumMember(typeEnum, "TIMESTAMP_NS", 22);
-	DefineEnumMember(typeEnum, "ENUM", 23);
-	DefineEnumMember(typeEnum, "LIST", 24);
-	DefineEnumMember(typeEnum, "STRUCT", 25);
-	DefineEnumMember(typeEnum, "MAP", 26);
-	DefineEnumMember(typeEnum, "ARRAY", 33);
-	DefineEnumMember(typeEnum, "UUID", 27);
-	DefineEnumMember(typeEnum, "UNION", 28);
-	DefineEnumMember(typeEnum, "BIT", 29);
-	DefineEnumMember(typeEnum, "TIME_TZ", 30);
-	DefineEnumMember(typeEnum, "TIMESTAMP_TZ", 31);
-  DefineEnumMember(typeEnum, "ANY", 34);
-  DefineEnumMember(typeEnum, "BIGNUM", 35);
-  DefineEnumMember(typeEnum, "SQLNULL", 36);
-  DefineEnumMember(typeEnum, "STRING_LITERAL", 37);
-  DefineEnumMember(typeEnum, "INTEGER_LITERAL", 38);
-  DefineEnumMember(typeEnum, "TIME_NS", 39);
-  DefineEnumMember(typeEnum, "GEOMETRY", 40);
-  DefineEnumMember(typeEnum, "VARIANT", 41);
-  return typeEnum;
-}
+#include "bindings_config.h"
+#include "conversion_helpers.h"
+#include "externals.h"
+#include "scalar_function_helpers.h"
+#include "promise_workers.h"
+#include "enums.h"
 
 // Addon
